requirements.txt
paramiko
paho-mqtt
redis